DAG
Here circles and squares correspond to latent and observed variables respectively. For a gene, either all samples or none of the samples are chosen after fitting the mixture model. Therefore the dashed arrows should not be in the DAG. If we would have chosen only samples with high correlations the dashed arrows are there and show that estimates of parameters modelling the relationship between disease severity and gene expression may be biased if we condition on the selection.